much C++ code removed (now in eclib-20250528) or refactored
