+++ Code/CurvesNj1728.cc
@@ -10,14 +10,14 @@
 
 // Example:
 //
-// $ ./CurvesNj 1728 1 100
+// $ ./CurvesNj1728 1 100
 // 32 [0,0,0,4,0]
 // 32 [0,0,0,-1,0]
 // 64 [0,0,0,1,0]
 // 64 [0,0,0,-4,0]
 
-
-#include "egros.h"
+#include <eclib/egros.h>
+#include "egros_cache.h"
 
 int main (int argc, char *argv[])
 {
@@ -55,12 +55,12 @@ int main (int argc, char *argv[])
     {
       bigint N(n);
       vector<bigint> suppN = pdivs(N);
-      if (test_conductor_j_1728(N, suppN))
+      if (is_N_possible_j_1728(N, suppN))
         {
           vector<CurveRed> Elist = get_egros_from_j_1728(pdivs(N));
-          for (auto Ei = Elist.begin(); Ei!=Elist.end(); ++Ei)
-            if (getconductor(*Ei) == N)
-              cout << N << " " << (Curve)(*Ei) <<endl;
+          for (auto E: Elist)
+            if (E.conductor() == N)
+              cout << N << " " << (Curve)E <<endl;
         }
     }
 }
